pyproject.toml
[project]
name = "vguild"
version = "0.1.0"
description = "Agent guild system for collaborative software task solving"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.12.0",
    "pydantic>=2.5.0",
    "PyYAML>=6.0",
    "rich>=13.0.0",
    "anthropic>=0.39.0",
]

[project.scripts]
vguild = "vguild.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vguild"]

[tool.ruff]
src = ["src"]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["B008"]

[tool.ruff.lint.isort]
known-first-party = ["vguild"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]
